changed copy_env script

+++ postbuild/scripts/copy_env.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 
+if [ -f ".env.local" ]; then
+  exit 0
+fi
+
 # Copy .env.example to .env.local
 cp .env.example .env.local
